Removed Watson STT, simplified TTS
